refactor: Moved config to their own folder and moved util functions to helper file
