+++ __tests__/prompts.test.js
@@ -1,16 +1,18 @@
-const prompts = require("../generators/app/src/prompts");
+const prompts = require("../generators/app/config/prompts").prompts;
+const licenses = require("../generators/app/config/prompts").licenses;
+const helper = require("../generators/app/src/helper")
 
 describe("prompts.js", () => {
 	describe("getLicenseValue", () => {
 		it("Should return the value for a given name", () => {
-			const name = prompts.licenses[0].name;
-			const expectedValue = prompts.licenses[0].value;
-			expect(prompts.getLicenseValue(name)).toEqual(expectedValue);
+			const name = licenses[0].name;
+			const expectedValue = licenses[0].value;
+			expect(helper.getLicenseValue(name)).toEqual(expectedValue);
 		});
 
 		it("Should return nothing when name unknown", () => {
 			const name = "unknown";
-			expect(prompts.getLicenseValue(name)).toBeUndefined();
+			expect(helper.getLicenseValue(name)).toBeUndefined();
 		});
 	});
 
@@ -26,7 +28,7 @@ describe("prompts.js", () => {
 				it("defined", () => {
 					expect.assertions(2);
 					const opts = { ...options, "role-name": "defined" };
-					const prom = prompts.promptBuilder(opts);
+					const prom = prompts(opts);
 
 					expect(prom[0].name).toEqual("roleName");
 					expect(prom[0].when).toBeFalsy();
@@ -35,7 +37,7 @@ describe("prompts.js", () => {
 				it("undefined", () => {
 					expect.assertions(2);
 					const opts = { ...options };
-					const prom = prompts.promptBuilder(opts);
+					const prom = prompts(opts);
 
 					expect(prom[0].name).toEqual("roleName");
 					expect(prom[0].when).toBeTruthy();
@@ -47,13 +49,13 @@ describe("prompts.js", () => {
 			describe("gitIncludeRepoUrl", () => {
 				it("true", () => {
 					response = { gitIncludeRepoUrl: true };
-					const prom = prompts.promptBuilder(options);
+					const prom = prompts(options);
 					expect(prom[6].when(response)).toEqual(response.gitIncludeRepoUrl);
 				});
 
 				it("false", () => {
 					response = { gitIncludeRepoUrl: false };
-					const prom = prompts.promptBuilder(options);
+					const prom = prompts(options);
 					expect(prom[6].when(response)).toEqual(response.gitIncludeRepoUrl);
 				});
 			});
@@ -64,7 +66,7 @@ describe("prompts.js", () => {
 				it("defined", () => {
 					expect.assertions(2);
 					const opts = { ...options, "include-molecule": "defined" };
-					const prom = prompts.promptBuilder(opts);
+					const prom = prompts(opts);
 
 					expect(prom[7].name).toEqual("includeMolecule");
 					expect(prom[7].when).toBeFalsy();
@@ -72,7 +74,7 @@ describe("prompts.js", () => {
 				it("undefined", () => {
 					expect.assertions(2);
 					const opts = { ...options };
-					const prom = prompts.promptBuilder(opts);
+					const prom = prompts(opts);
 
 					expect(prom[7].name).toEqual("includeMolecule");
 					expect(prom[7].when).toBeTruthy();
@@ -82,28 +84,28 @@ describe("prompts.js", () => {
 			it("include-molecule defined   | includeMolecule true", () => {
 				const response = { includeMolecule: true };
 				const opts = { ...options, "include-molecule": "defined" };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[8].when(response)).toBeTruthy();
 			});
 
 			it("include-molecule defined   | includeMolecule false", () => {
 				const response = { includeMolecule: false };
 				const opts = { ...options, "include-molecule": "defined" };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[8].when(response)).toBeTruthy();
 			});
 
 			it("include-molecule undefined | includeMolecule true", () => {
 				const response = { includeMolecule: true };
 				const opts = { ...options };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[8].when(response)).toBeTruthy();
 			});
 
 			it("include-molecule undefined | includeMolecule false", () => {
 				const response = { includeMolecule: false };
 				const opts = { ...options };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[8].when(response)).toBeFalsy();
 			});
 		});
@@ -113,7 +115,7 @@ describe("prompts.js", () => {
 				it("defined", () => {
 					expect.assertions(2);
 					const opts = { ...options, "include-meta": "defined" };
-					const prom = prompts.promptBuilder(opts);
+					const prom = prompts(opts);
 
 					expect(prom[9].name).toEqual("includeMeta");
 					expect(prom[9].when).toBeFalsy();
@@ -121,7 +123,7 @@ describe("prompts.js", () => {
 				it("undefined", () => {
 					expect.assertions(2);
 					const opts = { ...options };
-					const prom = prompts.promptBuilder(opts);
+					const prom = prompts(opts);
 
 					expect(prom[9].name).toEqual("includeMeta");
 					expect(prom[9].when).toBeTruthy();
@@ -131,28 +133,28 @@ describe("prompts.js", () => {
       it("include-meta defined   | includeMeta true", () => {
 				const response = { includeMeta: true };
 				const opts = { ...options, "include-meta": "defined" };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[10].when(response)).toBeTruthy();
 			});
 
 			it("include-meta defined   | includeMeta false", () => {
 				const response = { includeMeta: false };
 				const opts = { ...options, "include-meta": "defined" };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[10].when(response)).toBeTruthy();
 			});
 
 			it("include-meta undefined | includeMeta true", () => {
 				const response = { includeMeta: true };
 				const opts = { ...options };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[10].when(response)).toBeTruthy();
 			});
 
 			it("include-meta undefined | includeMeta false", () => {
 				const response = { includeMeta: false };
 				const opts = { ...options };
-				const prom = prompts.promptBuilder(opts);
+				const prom = prompts(opts);
 				expect(prom[10].when(response)).toBeFalsy();
 			});
 		});
