+++ generators/app/config/arguments.js
@@ -11,10 +11,4 @@ const args = [
 	}
 ];
 
-const generateArgs = generator => {
-	for (arg of args) {
-		generator.argument(arg.name, arg.options);
-	}
-};
-
-module.exports = { generateArgs, args };
+module.exports = { args };
