+++ generators/app/src/package-builder.js
@@ -1,25 +1,10 @@
-const getLicenseValue = require("./prompts").getLicenseValue;
+const pkgTemplate = require("../config/package-json").pkgTemplate;
 
 const buildPackageJSON = props => {
-	let pkgTemplate = {
-		name: props.roleName,
-		version: "1.0.0",
-		main: "index.js",
-		repository: props.gitIncludeRepoUrl ? props.gitRepoUrl : "",
-		description: props.description,
-		author: props.gitAuthorName,
-		license: getLicenseValue(props.license),
-		dependencies: {},
-		scripts: {
-			venv: "virtualenv venv",
-			"venv-activate": "source venv/bin/activate",
-			requirements: "pip install -r requirements.txt"
-		}
-	};
-
+	const newTemplate = { ...pkgTemplate(props) };
 	if (props.includeMolecule) {
-		pkgTemplate.scripts = {
-			...pkgTemplate.scripts,
+		newTemplate.scripts = {
+			...newTemplate.scripts,
 			...{
 				check: "molecule check",
 				converge: "molecule converge",
@@ -40,7 +25,7 @@ const buildPackageJSON = props => {
 		};
 	}
 
-	return pkgTemplate;
+	return newTemplate;
 };
 
 module.exports = {
